+++ Providers/AspNet/Revo.AspNet/App_Start/WebApiConfig.cs
@@ -1,10 +1,4 @@
-﻿using System.Web.Helpers;
-using System.Web.Http;
-using Microsoft.AspNet.OData.Extensions;
-using Newtonsoft.Json.Converters;
-using Newtonsoft.Json.Serialization;
-using Revo.AspNet.Globalization;
-using Revo.AspNet.IO;
+﻿using System.Web.Http;
 using Revo.AspNet.Web;
 using Revo.Core.Configuration;