+++ Revo.Core/Globalization/ITranslatable.cs
@@ -1,9 +1,0 @@
-﻿namespace Revo.Core.Globalization
-{
-    public interface ITranslatable
-    {
-        string Code { get; }
-        string Name { get; }
-        string Culture { get; set; }
-    }
-}