+++ Revo.Domain/Core/IClassifier.cs
@@ -1,7 +1,0 @@
-﻿namespace Revo.Domain.Core
-{
-    public interface IClassifier
-    {
-        string Code { get; }
-    }
-}