+++ Revo.Infrastructure/Globalization/ILocaleManager.cs
@@ -1,9 +1,0 @@
-﻿using System.Collections.Generic;
-
-namespace Revo.Infrastructure.Globalization
-{
-    public interface ILocaleManager
-    {
-        IReadOnlyDictionary<string, Locale> Locales { get; }
-    }
-}