+++ Revo.Infrastructure/Globalization/MessageRepositoryReloadedEvent.cs
@@ -1,9 +1,0 @@
-﻿using Revo.Core.Events;
-using Revo.Domain.Events;
-
-namespace Revo.Infrastructure.Globalization
-{
-    public class MessageRepositoryReloadedEvent : IEvent
-    {
-    }
-}