+++ Revo.Infrastructure/Globalization/Messages/Database/DbMessageCacheReloadedEvent.cs
@@ -1,9 +1,0 @@
-﻿using Revo.Core.Events;
-using Revo.Domain.Events;
-
-namespace Revo.Infrastructure.Globalization.Messages.Database
-{
-    public class DbMessageCacheReloadedEvent : IEvent
-    {
-    }
-}