+++ Revo.Infrastructure/Globalization/Messages/Database/IDbMessageLoader.cs
@@ -1,8 +1,0 @@
-﻿namespace Revo.Infrastructure.Globalization.Messages.Database
-{
-    public interface IDbMessageLoader
-    {
-        void EnsureLoaded();
-        void Reload();
-    }
-}