+++ Revo.Infrastructure/Globalization/Messages/Database/LocalizationMessageDeletedEvent.cs
@@ -1,8 +1,0 @@
-﻿using Revo.Domain.Events;
-
-namespace Revo.Infrastructure.Globalization.Messages.Database
-{
-    public class LocalizationMessageDeletedEvent : DomainAggregateEvent
-    {
-    }
-}